requirements.txt
fastapi
uvicorn
pandas
openpyxl
scikit-learn
sentence-transformers
openai>=1.0.0
python-dotenv>=1.0.0
